requirements.txt
strands-agents[openai]
requests
beautifulsoup4
pdfplumber
pydantic

# Optional
# chromadb
# sentence-transformers

